Check the Windows certificate store for custom CAs
Sets security.enterprise_roots.enabled to true

This will hopefully fix the vast majority of certificate errors from
custom CAs on Windows. While it means that people won't receive any
indication of a MITM, that's not really our job, and their institution
and/or AV software and/or malware is already intercepting other secure
traffic on their system.

This has no effect on macOS, where the pref wasn't supported until
Firefox 63, or Linux, which is complicated [1].

Further reading:

https://mike.kaply.com/2016/09/01/upcoming-changes-to-root-certificates-in-firefox-on-windows/

Recent Mozilla Shield experiment turning this pref on, with more details
of what exactly it does:

https://bugzilla.mozilla.org/show_bug.cgi?id=1533397

[1] https://wiki.mozilla.org/CA/AddRootToFirefox#Linux